// HRM/ClientApp/src/app/modules/staff/models/staff-model.ts
export class StaffModel {
    id: number;
    firstName: string;
    lastName: string;
    dateOfBirth: Date;
    sex: number;
    address: string;
    email: string;
    phone: string;
    contractDate: Date;
    departmentId: number;
    positionId: number;
    imageLink: string;
    createDate: Date;
    createId: number;
    updateDate: Date;
    updateId: number;
    personalId: string;
}